ref: Move and organize several components
- Moved several files from "lib/" -> "components/custom/ui/"
  - `motion-anim.tsx`
  - `ThemeToggle.tsx`
  - `FocusableCard.tsx`
- Moved several files from "lib/" -> "components/layout/"
  - `Article.tsx`
  - `Layout.tsx`
