+++ src/layouts/HomeLayout.tsx
@@ -4,9 +4,9 @@ import { unified } from 'unified';
 import { CardContent } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { cn } from "@/lib/utils";
-import { Section } from "@/lib/motion-anim";
-import Article from "@/lib/Article";
-import FocusableCard from "@/lib/FocusableCard";
+import { Section } from "@/components/custom/ui/motion-anim";
+import Article from "@/components/layout/Article";
+import FocusableCard from "@/components/custom/ui/FocusableCard";
 import { unify } from "@/lib/markparser";
 import { remarkSlugFromHeading, type Heading } from "@/lib/plugins/remark/remark-slug-from-heading";
 import { extractSections } from '@/utils/article';
